requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
click==8.3.1
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
distro==1.9.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jiter==0.12.0
msgpack==1.1.2
openai==2.15.0
packaging==25.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyOpenSSL==25.3.0
redis==7.1.0
service-identity==24.2.0
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
uvicorn==0.40.0
zope.interface==8.2
